trying to tease apart the decryption part that should be in LicenseDownloader
