+++ src/KeyHub.Client/LicenseVerifier.cs
@@ -200,7 +200,7 @@ namespace KeyHub.Client
         /// <param name="featureDisplayName"></param>
         public void NotifyUse(string domain, Guid feature)
         {
-            domain = NormalizeDomain(domain);
+            domain = DomainUtility.NormalizeDomain(domain);
 
             FeatureState state;
             //Find out what the state of this feature is in the local cache, or initialize to Pending if unknown.
@@ -234,16 +234,6 @@ namespace KeyHub.Client
             }
 
             PingBackgroundVerification();
-        }
-
-        public string NormalizeDomain(string domain)
-        {
-            //lowercase
-            domain = domain.ToLowerInvariant();
-            //Strip www prefix off.
-            if (domain.StartsWith("www.")) domain = domain.Substring(4);
-
-            return domain;
         }
 
         private DateTime lastScheduledVerification = DateTime.MinValue;
@@ -280,7 +270,7 @@ namespace KeyHub.Client
                     ILicenseStore s = config.Plugins.GetOrInstall<ILicenseStore>(new LicenseStore());
                     //Get all encrypted licenses and 
                     // Decrypt them, grouping by normalized domain name
-                    var licenses = RemovedExpired(DecryptAll(s.GetLicenses()));
+                    var licenses = RemovedExpired(new LicenseVerifierBase().DecryptAll(PublicKeyXml, s.GetLicenses()));
                     
                     //Fix all possible pendingFeatures using existing license data.
                     UpdateFeatureStatus(licenses);
@@ -296,18 +286,18 @@ namespace KeyHub.Client
                         string appStr = config.get("licenses.auto.appId", null);
                         if (!string.IsNullOrEmpty(appStr))
                         {
-                            requestLicenses = new LicenseDownloader().RequestLicenses(LicensingUrl, new Guid(appStr), pendingFeatures);                            
-                        }
-                        
-                        var newLicenses = DecryptAll(requestLicenses);
+                            requestLicenses = new LicenseDownloader().RequestLicenses(LicensingUrl, new Guid(appStr), pendingFeatures);
 
-                        if (newLicenses.Count > 0)
-                        {
-                            //Update cached feature statuses
-                            UpdateFeatureStatus(newLicenses);
+                            var newLicenses = new LicenseVerifierBase().DecryptAll(PublicKeyXml, requestLicenses);
 
-                            s.SetLicenses(ExportLicenses(Merge(licenses, newLicenses)));
-                        }
+                            if (newLicenses.Count > 0)
+                            {
+                                //Update cached feature statuses
+                                UpdateFeatureStatus(newLicenses);
+
+                                s.SetLicenses(ExportLicenses(Merge(licenses, newLicenses)));
+                            }
+                        }                        
                     }
                 }
                 finally
@@ -416,25 +406,6 @@ namespace KeyHub.Client
             }
         }
 
-        private Dictionary<string, List<DomainLicense>> DecryptAll(ICollection<byte[]> encrypted)
-        {
-            var licenses = new Dictionary<string, List<DomainLicense>>(StringComparer.OrdinalIgnoreCase);
-
-            foreach (byte[] data in encrypted)
-            {
-                var d = new DomainLicense(data, this);
-                string domain = NormalizeDomain(d.Domain);
-                List<DomainLicense> forDomain;
-                if (!licenses.TryGetValue(domain, out forDomain))
-                {
-                    forDomain = new List<DomainLicense>();
-                    licenses[domain] = forDomain;
-                }
-                forDomain.Add(d);
-            }
-            return licenses;
-        }
-
         /// <summary>
         /// Generates a key 2048-bit keypair and returns the xml fragment containing it
         /// </summary>
@@ -455,107 +426,6 @@ namespace KeyHub.Client
             {
                 r.FromXmlString(pair);
                 return r.ToXmlString(false);
-            }
-        }
-
-        private static object decryptionLock = new object();
-        private static RSACryptoServiceProvider crypto;
-
-
-        internal string Decrypt(byte[] data)
-        {
-            lock (decryptionLock)
-            {
-                if (crypto == null)
-                {
-                    crypto = new RSACryptoServiceProvider(2048);
-                    crypto.FromXmlString(PublicKeyXml);
-                }
-
-                byte[] decrypted = crypto.Decrypt(data, false);
-                return System.Text.UTF8Encoding.UTF8.GetString(decrypted);
-            }
-        }
-
-
-        class DomainLicense
-        {
-
-            internal DomainLicense(byte[] encryptedLicense, LicenseVerifier v)
-            {
-                Decrypted = v.Decrypt(encryptedLicense);
-                Encrypted = encryptedLicense;
-                string[] lines = Decrypted.Split('\n');
-                foreach (string l in lines)
-                {
-                    int colon = l.IndexOf(':');
-                    if (colon < 1) continue;
-                    string key = l.Substring(0, colon).Trim().ToLowerInvariant();
-                    string value = l.Substring(colon + 1).Trim();
-
-                    switch (key)
-                    {
-                        case "domain": Domain = value; break;
-                        case "owner": OwnerName = value; break;
-                        case "issued": Issued = DateTime.Parse(value); break;
-                        case "expires": Expires = DateTime.Parse(value); break;
-                        case "features":
-                            List<Guid> ids = new List<Guid>();
-                            string[] parts = value.Split(new char[] { ',' }, StringSplitOptions.RemoveEmptyEntries);
-                            foreach (string p in parts)
-                            {
-                                ids.Add(new Guid(p));
-                            }
-                            Features = ids;
-                            break;
-                    }
-
-                }
-            }
-
-            internal string Decrypted { get; private set; }
-            internal byte[] Encrypted { get; private set; }
-            internal string Domain { get; private set; }
-            internal string OwnerName { get; private set; }
-            internal DateTime Issued { get; private set; }
-            internal DateTime Expires { get; private set; }
-            internal IList<Guid> Features { get; private set; }
-
-            internal string Join(ICollection<Guid> items)
-            {
-                StringBuilder sb = new StringBuilder();
-                foreach (Guid g in items)
-                    sb.Append(g.ToString() + ",");
-                return sb.ToString().TrimEnd(',');
-            }
-
-            internal string SerializeUnencrypted()
-            {
-                return "Domain: " + Domain.Replace('\n', ' ') + "\n" +
-                        "OwnerName: " + OwnerName.Replace('\n', ' ') + "\n" +
-                        "Issued: " + Issued.ToString() + "\n" +
-                        "Expires: " + Expires.ToString() + "\n" +
-                        "Features: " + Join(Features) + "\n";
-            }
-
-            internal byte[] SerializeAndEncrypt(string xmlKeyPair)
-            {
-                using (var r = new RSACryptoServiceProvider(2048))
-                {
-                    r.FromXmlString(xmlKeyPair);
-                    return r.Encrypt(UTF8Encoding.UTF8.GetBytes(SerializeUnencrypted()), false);
-                }
-            }
-            /// <summary>
-            /// Returns a human readable, single-line description of the license
-            /// </summary>
-            /// <returns></returns>
-            internal string GetShortDescription()
-            {
-                StringBuilder sb = new StringBuilder(OwnerName + " - " + Domain + " - " + Issued.ToString() + " - " + Expires.ToString() + " - ");
-                foreach (var id in Features)
-                    sb.Append(GetFriendlyName(id) + " ");
-                return sb.ToString().TrimEnd();
             }
         }
